Lights now go in sequence without interrupting RFID reads
- The NeoPixels on the prototype suit now turn on sequentially, without
delays and without interrupting the RFID reader's polling. We did this
by using millis() and comparing the current millis() read to the last
one, and if it's bigger than an interval (like 250 milliseconds), it
turns on the next light and turns off the previous light.
- Renamed SuitNew to Suit, and the previous "Suit" to SuitOld
- The methods on the suit's end have now been split up into separate
files within the sketch to make it more readable. The sketch now
contains: Suit (setup and loop), XBee, LEDs, RFID.